outsource gef into a separate script
